Adjust `ProductionDataTransferFeed` entity migration
- Renamed `remaining_value` to `value` and `remaining_tonnage` to `tonnage` in `production_data_transfer_feed` table.
- Updated EF Core model snapshot to reflect these changes.

+++ backend/src/FarmsManager.Application/Commands/ProductionData/RemainingFeed/AddProductionDataRemainingFeedCommand.cs
@@ -20,7 +20,7 @@ public record AddProductionDataRemainingFeedCommand : IRequest<EmptyBaseResponse
     public Guid HenhouseId { get; init; }
     public Guid CycleId { get; init; }
     public string FeedName { get; init; }
-    public int RemainingTonnage { get; init; }
+    public decimal RemainingTonnage { get; init; }
     public decimal RemainingValue { get; init; }
 }
